requirements.txt
sqlalchemy~=2.0.27
pydantic~=2.5.3
fastapi~=0.103.0
requests~=2.31.0
uvicorn~=0.20.0
python-jose~=3.3.0
passlib~=1.7.4
python-multipart~=0.0.9
bcrypt==4.0.1